Move out classes from participant-state package object
Having the classes specified in the package object makes the
unusable from Java.
